Move hooks into hooklist
