+++ packages/core/src/arena.rs
@@ -110,7 +110,7 @@ impl VirtualDom {
 
         // Drop all the hooks once the children are dropped
         // this means we'll drop hooks bottom-up
-        scope.hooks.get_mut().clear();
+        scope.hooks.clear();
 
         // Drop all the futures once the hooks are dropped
         for task_id in scope.spawned_tasks.borrow_mut().drain() {
